added RR and handler

+++ scheduler.c
@@ -40,7 +40,7 @@ void connectWithGenerator();
 void finishedPhandler(int signum);
 void sigtermhandler(int signum);
 void RRScheduler(int quantum);
-void processTerminated(int signum);
+// void processTerminated(int signum);
 process *initProcess();
 void outputFileinit();
 void outputFile();
@@ -58,7 +58,7 @@ int main(int argc, char *argv[])
 
   signal(SIGTERM, sigtermhandler);    // to free the allocated memory
   signal(SIGUSR1, finishedPhandler);  // to recieve that a process has finished its execution
-  signal(SIGUSR2, processTerminated); // to handle the termination of a process
+  // signal(SIGUSR2, processTerminated); // to handle the termination of a process
   algorithm = atoi(argv[1]);
   processCount = atoi(argv[2]);
   finishedQueue = createQueue();
@@ -262,14 +262,22 @@ void sigtermhandler(int signum)
 
 void finishedPhandler(int signum)
 {
-  process *finishedprocess = NULL;
+  if (algorithm != 3)
+  {
+    process *finishedprocess = NULL;
+    runningProcess = NULL;
+    finishedprocess = PQdequeue(PQ);
+    printf("Process ID = %d Fininshed at time = %d\n", finishedprocess->id, getClk());
 
-  finishedprocess = PQdequeue(PQ);
-  runningProcess = NULL;
-  printf("Process ID = %d Fininshed at time = %d\n", finishedprocess->id, getClk());
-
-  free(finishedprocess);
-  processCount--;
+    free(finishedprocess);
+    processCount--;
+  }
+  else
+  {
+    // when a process fnishes it should notify the scheduler on termination, the scheduler does NOT terminate the process.
+    // the scheduler should remove the process from the queue and free its memory
+    flag = 1;
+  }
   signal(SIGUSR1, finishedPhandler);
 }
 
@@ -277,45 +285,69 @@ void finishedPhandler(int signum)
 
 void RRScheduler(int quantum)
 {
-  // when a process fnishes it should notify the scheduler on termination, the scheduler does NOT terminate the process.
-
-  runningProcess = dequeue(Q);
-
-  int reamingtime = runningProcess->runningtime;
-
-  while (!isEmpty(Q))
+  Q = createQueue();
+  printf("Queue created with quantum = %d\n", quantum);
+  int counter = 0;
+  while (true)
   {
-    int runtime;
-    if (quantum < reamingtime)
-      runtime = quantum;
-    else
-      runtime = reamingtime;
-
-    kill(runningProcess->realPid, SIGCONT);
-    sleep(runtime);
-
-    if (flag)
+    if (!isEmpty(Q))
     {
-      kill(runningProcess->realPid, SIGSTOP);
-      normalQenqueue(Q, runningProcess);
+      runningProcess = dequeue(Q);
+
+      int remainingtime = runningProcess->remainingtime;
+      int runtime;
+      
+      if (quantum < remainingtime)
+        runtime = quantum;
+      else
+        runtime = remainingtime;
+
+      kill(runningProcess->realPid, SIGCONT);
+      sleep(runtime);
     }
-    else
+
+    // This loop checks for the incoming processes, if there is no incoming processes, it will break and continue running the current process
+    if (processCount > 0)
     {
-      free(runningProcess);
-      flag = 1;
+      // int lastID = -1;
+      while ((msgrcv(sch_msgq_id, &SCH_message, sizeof(SCH_message.arrivedProcess), getpid(), IPC_NOWAIT)) != -1)
+      {
+        // if (lastID == SCH_message.arrivedProcess.id)
+        //   continue;
+
+        printf("Received\n");
+        process *newprocess = initProcess();
+        normalQenqueue(Q, newprocess);
+
+        // lastID = newprocess->id;
+      }
     }
+
+    if (runningProcess)
+    {
+      // when a process fnishes it should notify the scheduler on termination, the scheduler does NOT terminate the process.
+      if (flag)
+      {
+        kill(runningProcess->realPid, SIGSTOP);
+        normalQenqueue(Q, runningProcess);
+      }
+      else
+      {
+        free(runningProcess);
+        processCount--;
+        flag = 1;
+        // printf("Done cleaning\n");
+      }
+    }
+
+    if (processCount == 0)
+      break;
   }
+
+  free(Q);
 }
 
 ///////////////////////////////////////////
-
-void processTerminated(int signum)
-{
-  // when a process fnishes it should notify the scheduler on termination, the scheduler does NOT terminate the process.
-  // the scheduler should remove the process from the queue and free its memory
-
-  flag = 0;
-}
 
 process *initProcess()
 {
@@ -375,3 +407,14 @@ void printProcessState(FILE *schedulerLog, int time, int processID, char *state,
   // fprintf(schedulerLog, "#At time %d process %d state %s arr %d total %d remain %d wait %d\n",
   //         time, processID, state, arrivalTime, totalRunningTime, remainingTime, waitingTime);
 }
+
+
+///////////////////////////////////////////
+
+// void processTerminated(int signum)
+// {
+//   // when a process fnishes it should notify the scheduler on termination, the scheduler does NOT terminate the process.
+//   // the scheduler should remove the process from the queue and free its memory
+
+//   flag = 0;
+// }
